feat: Type of Joins.
Exercises with Inner join, which joins tables according to one or more condisitons.

Exercises with Group join, it does not exists in SQL, but is used to make groups, is more likely to use aggregation functions.

Finally, exercises with cross join, wich is the combination of the tables, it allows us to know all the possible combinations, in this case between two tables, Courses and Authors.